requirements.txt
requests==2.18.4
sendgrid-python==5.1.0